+++ plugins/kalabox-core/lib/tasks/start.js
@@ -13,6 +13,12 @@ module.exports = function(kbox) {
       kbox.tasks.add(function(task) {
         task.path = [app.name, 'start'];
         task.category = 'appAction';
+        task.options.push({
+          name: 'force',
+          alias: 'f',
+          kind: 'boolean',
+          description: 'Force start even if already running.'
+        });
         task.description = 'Start an installed kbox application.';
         task.func = function() {
 
@@ -20,12 +26,15 @@ module.exports = function(kbox) {
           var format = require('util').format;
           var log = kbox.core.log;
 
+          // Get options
+          var options = this.options || {};
+
           // Check to see if app is already running
           return kbox.app.isRunning(app)
 
           // Start if not running, otherwise inform user
           .then(function(isRunning) {
-            if (!isRunning) {
+            if (!isRunning || options.force) {
               return kbox.app.start(app)
               .then(function() {
                 console.log(kbox.art.appStart(app));